+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- CinemachineFollowAttacher component
 ### Changed
 - Improve First Region creation. Check if main camera is perspective or orthographic
 - Improve Create Regions GUI. Inspector and Overlay Scene show the same extra GUI helpers